HW6+Optional+fix

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -52,12 +52,14 @@ public class JdbcUserRepository implements UserRepository {
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             user.setId(newKey.intValue());
-        } else if (namedParameterJdbcTemplate.update(
-                "UPDATE users SET name=:name, email=:email, password=:password, " +
-                        "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
-            return null;
+        } else {
+            if (namedParameterJdbcTemplate.update(
+                    "UPDATE users SET name=:name, email=:email, password=:password, " +
+                            "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
+                return null;
+            }
+            deleteAllRoles(user.getId());
         }
-        deleteAllRoles(user.getId());
         addRole(user);
         return user;
     }
@@ -70,16 +72,12 @@ public class JdbcUserRepository implements UserRepository {
 
     @Override
     public User get(int id) {
-        List<User> users = jdbcTemplate.query("SELECT * FROM users WHERE id=?", ROW_MAPPER, id);
-        Collections.singleton(DataAccessUtils.singleResult(users)).forEach(user->setRoles(user));
-        return DataAccessUtils.singleResult(users);
+        return setRoles(jdbcTemplate.query("SELECT * FROM users WHERE id=?", ROW_MAPPER, id));
     }
 
     @Override
     public User getByEmail(String email) {
-        List<User> users = jdbcTemplate.query("SELECT * FROM users WHERE email=?", ROW_MAPPER, email);
-        Collections.singleton(DataAccessUtils.singleResult(users)).forEach(user->setRoles(user));
-        return DataAccessUtils.singleResult(users);
+        return setRoles(jdbcTemplate.query("SELECT * FROM users WHERE email=?", ROW_MAPPER, email));
     }
 
     @Override
@@ -111,11 +109,13 @@ public class JdbcUserRepository implements UserRepository {
         jdbcTemplate.update(query, userId);
     }
 
-    private User setRoles(User user) {
+    private User setRoles(List<User> users) {
+        User user = DataAccessUtils.singleResult(users);
         if (user !=null) {
            List<Role> roles = jdbcTemplate.query("SELECT * FROM user_roles WHERE user_id=?", (rs, rowNum) -> Role.valueOf(rs.getString("role")), user.getId());
            user.setRoles(roles);
         }
+
         return user;
     }
 }
